Move location of durable message cleaner task
