+++ datastores/standarddb/db/tasks.php
@@ -1,7 +1,7 @@
 <?php
 $tasks = [
     [
-        'classname' => 'mbdatastore_standarddb\durable_message_cleaner',
+        'classname' => 'mbdatastore_standarddb\task\durable_message_cleaner',
         'blocking' => 0,
         'minute' => 'R',
         'hour' => '*/3',
