+++ datastores/standarddb/version.php
@@ -3,6 +3,6 @@ defined('MOODLE_INTERNAL') || die();
 
 /** @var stdClass $plugin */
 $plugin->component = 'mbdatastore_standarddb';
-$plugin->version = 2023060500;
+$plugin->version = 2023090400;
 $plugin->requires = 2020061505;
 $plugin->maturity = MATURITY_ALPHA;
